# Update PropertyValues in TriggerBase/FrameworkElementFactory via reference