--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/DataTrigger.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/DataTrigger.cs
@@ -162,7 +162,7 @@ namespace System.Windows
             // Set Condition for all data triggers
             for (int i = 0; i < PropertyValues.Count; i++)
             {
-                PropertyValue propertyValue = PropertyValues[i];
+                ref PropertyValue propertyValue = ref PropertyValues.GetEntryAtRef(i);
 
                 propertyValue.Conditions = TriggerConditions;
                 switch (propertyValue.ValueType)
@@ -176,9 +176,6 @@ namespace System.Windows
                     default:
                         throw new InvalidOperationException(SR.Format(SR.UnexpectedValueTypeForDataTrigger, propertyValue.ValueType));
                 }
-
-                // Put back modified struct
-                PropertyValues[i] = propertyValue;
             }
 
             base.Seal();