--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/MultiDataTrigger.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/MultiDataTrigger.cs
@@ -124,7 +124,7 @@ namespace System.Windows
             // Set conditions array for all property triggers
             for (int i = 0; i < PropertyValues.Count; ++i)
             {
-                PropertyValue propertyValue = PropertyValues[i];
+                ref PropertyValue propertyValue = ref PropertyValues.GetEntryAtRef(i);
                 propertyValue.Conditions = TriggerConditions;
                 switch (propertyValue.ValueType)
                 {
@@ -137,9 +137,6 @@ namespace System.Windows
                     default:
                         throw new InvalidOperationException(SR.Format(SR.UnexpectedValueTypeForDataTrigger, propertyValue.ValueType));
                 }
-
-                // Put back modified struct
-                PropertyValues[i] = propertyValue;
             }
 
             base.Seal();