--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Trigger.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Trigger.cs
@@ -224,10 +224,8 @@ namespace System.Windows
             // Set Condition for all property triggers
             for (int i = 0; i < PropertyValues.Count; i++)
             {
-                PropertyValue propertyValue = PropertyValues[i];
+                ref PropertyValue propertyValue = ref PropertyValues.GetEntryAtRef(i);
                 propertyValue.Conditions = TriggerConditions;
-                // Put back modified struct
-                PropertyValues[i] = propertyValue;
             }
 
             base.Seal();