--- a/src/Microsoft.DotNet.Wpf/src/Shared/MS/Utility/FrugalList.cs
+++ b/src/Microsoft.DotNet.Wpf/src/Shared/MS/Utility/FrugalList.cs
@@ -141,6 +141,8 @@ namespace MS.Utility
         /// </summary>
         public abstract T EntryAt(int index);
 
+        public abstract ref T EntryAtRef(int index);
+
         /// <summary>
         /// Promotes the values in the current store to the next larger
         /// and more complex storage model.
@@ -318,6 +320,11 @@ namespace MS.Utility
         public override T EntryAt(int index)
         {
             return _loneEntry;
+        }
+
+        public override ref T EntryAtRef(int index)
+        {
+            return ref _loneEntry;
         }
 
         public override void Promote(FrugalListBase<T> oldList)
@@ -564,6 +571,24 @@ namespace MS.Utility
 
                 case 2:
                     return _entry2;
+
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(index));
+            }
+        }
+
+        public override ref T EntryAtRef(int index)
+        {
+            switch (index)
+            {
+                case 0:
+                    return ref _entry0;
+
+                case 1:
+                    return ref _entry1;
+
+                case 2:
+                    return ref _entry2;
 
                 default:
                     throw new ArgumentOutOfRangeException(nameof(index));
@@ -1019,6 +1044,33 @@ namespace MS.Utility
             }
         }
 
+        public override ref T EntryAtRef(int index)
+        {
+            switch (index)
+            {
+                case 0:
+                    return ref _entry0;
+
+                case 1:
+                    return ref _entry1;
+
+                case 2:
+                    return ref _entry2;
+
+                case 3:
+                    return ref _entry3;
+
+                case 4:
+                    return ref _entry4;
+
+                case 5:
+                    return ref _entry5;
+
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(index));
+            }
+        }
+
         public override void Promote(FrugalListBase<T> oldList)
         {
             int oldCount = oldList.Count;
@@ -1390,6 +1442,11 @@ namespace MS.Utility
         public override T EntryAt(int index)
         {
             return _entries[index];
+        }
+
+        public override ref T EntryAtRef(int index)
+        {
+            return ref _entries[index];
         }
 
         public override void Promote(FrugalListBase<T> oldList)
@@ -2071,6 +2128,16 @@ namespace MS.Utility
                 }
                 throw new ArgumentOutOfRangeException(nameof(index));
             }
+        }
+
+        public ref T GetEntryAtRef(int index)
+        {
+            // If no entry, default(T) is returned
+            if ((null != _listStore) && ((index < _listStore.Count) && (index >= 0)))
+            {
+                return ref _listStore.EntryAtRef(index);
+            }
+            throw new ArgumentOutOfRangeException(nameof(index));
         }
 
         public int Add(T value)