## Description

Adds ability to retrieve elements via `ref` from `FrugalStructList<T>` and the underlying storage formats, allowing for future optimizations. I intentionally did not modify the indexer as this can be done easily later on, so there's a new method for it now, which will allow for incremental optimizations in future on another types, where it makes sense and is safe to do so.

This allows us to update items via reference, which is particularly useful in case of `PropertyValues` used in `TriggerBase` and `FrameworkElementFactory` since it is a struct that needs to be copied everytime on update back and forth.

It is particularly efficient optimization here since the struct is full of managed pointers and ain't blittable, which takes considerable time to track back and forth for a simple update on one property.

I've also updated the getter/setter on `FrugalStructList<T>` for a minor speed boost, since it is used in almost every WPF collection, every branch counts. Nothing extra new, this is used in `List<T>` by default for several releases.

### Original getter/setter prolog:

```assembly
mov rcx, [rcx+8]
test rcx, rcx      
je short L0021 
cmp edx, [rcx+8]     
jge short L0021
test edx, edx
jl short L0021
```

### New getter/setter prolog:

```assembly
mov rcx, [rcx+8]
test rcx, rcx    
je short L001d 
cmp edx, [rcx+8] 
jae short L001d
```

### Updating PropertyValues before/after

| Method          |  Mean [ns] | Error [ns] | StdDev [ns] | Code Size [B] | Allocated [B] |
|---------------- |----------:|-----------:|------------:|--------------:|--------------:|
| Original | 459.11 ns |   2.278 ns |    2.131 ns |       1,072 B |             - |
| PR__EDIT|  58.45 ns |   0.464 ns |    0.434 ns |         469 B |             - |

<details><summary>Minimal benchmark code</summary>

```csharp
[Benchmark]
public static FrugalStructList<PropertyValue> PropertyValues = new();
public static FrugalStructList<PropertyValue> PropertyValues2 = new();

public static IEnumerable<TriggerCondition[]> KeysToTest
{
    get
    {
        yield return new TriggerCondition[] { new TriggerCondition() {  SourceChildIndex = 4}, new TriggerCondition() { SourceChildIndex = 1 }, new TriggerCondition() { SourceChildIndex = 3 } };
    }
}

[GlobalSetup]
public static void Setup()
{
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });

  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
  PropertyValues2.Add(new PropertyValue() { ChildName = "abcd", Conditions = new TriggerCondition[5], Property = TextBox.TextProperty, ValueInternal = 45, ValueType = PropertyValueType.Trigger });
}

[ArgumentsSource(nameof(KeysToTest))]
public void Original(TriggerCondition[] triggerCondition)
{
    // Set Condition for all data triggers
    for (int i = 0; i < PropertyValues.Count; i++)
    {
        PropertyValue propertyValue = PropertyValues[i];

        propertyValue.Conditions = triggerCondition;
        switch (propertyValue.ValueType)
        {
            case PropertyValueType.Trigger:
                propertyValue.ValueType = PropertyValueType.DataTrigger;
                break;
            case PropertyValueType.DataTrigger:
                propertyValue.ValueType = PropertyValueType.Trigger;
                break;
            case PropertyValueType.PropertyTriggerResource:
                propertyValue.ValueType = PropertyValueType.DataTriggerResource;
                break;
            default:
                throw new InvalidOperationException("fail");
        }

        // Put back modified struct
        PropertyValues[i] = propertyValue;
    }
}

[Benchmark]
[ArgumentsSource(nameof(KeysToTest))]
public void PR__EDIT(TriggerCondition[] triggerCondition)
{
    // Set Condition for all data triggers
    for (int i = 0; i < PropertyValues2.Count; i++)
    {
        ref PropertyValue propertyValue = ref PropertyValues2.GetEntryAtRef(i);

        propertyValue.Conditions = triggerCondition;
        switch (propertyValue.ValueType)
        {
            case PropertyValueType.Trigger:
                propertyValue.ValueType = PropertyValueType.DataTrigger;
                break;
            case PropertyValueType.DataTrigger:
                propertyValue.ValueType = PropertyValueType.Trigger;
                break;
            case PropertyValueType.PropertyTriggerResource:
                propertyValue.ValueType = PropertyValueType.DataTriggerResource;
                break;
            default:
                throw new InvalidOperationException("fail");
        }
    }
}
```

</details>

## Customer Impact

Increased startup performance mostly, but also when dynamically creating triggers at runtime and obviously everything related to ContentPresenter, Validation, ItemsControl templates etc. I could observe about several milliseconds improvement with a simple app with a few controls with ScrollViewer (ListBox) etc.

## Regression

No.

## Testing

Local build, sample apps testing with styles/triggers to check they're applied property.

## Risk

Low-to-medium, reviewers shall check that the `ref` is not forwarded/used anywhere while the `FrugalStructList<T>` could be modified (this really means when Grow or for example Remove happens), R/W operations via indexer do not matter as they do not modify the backing store in size/copy it.